+++ main/playlist_manager/playlist_manager_statistics.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//16/09/26
+//17/09/26
 
 /* exported _listStatistics */
 
@@ -155,7 +155,7 @@ function _listStatistics(x, y, w, h, bEnabled = false, config = {}) {
 		// helper
 		const createMenuOption = createMenuOptionParent.bind(this, menu);
 		// Header
-		menu.newEntry({ entryText: this.title, flags: MF_GRAYED });
+		menu.newEntry({ entryText: this.title.key.cut(50), flags: MF_GRAYED });
 		menu.newSeparator();
 		{	// Data
 			const subMenu = menu.newMenu('Data');